renamed full_transaction_processor
