+++ zingolib/src/blaze.rs
@@ -2,8 +2,8 @@
 
 pub(super) mod block_management_reorg_detection;
 pub(super) mod fetch_compact_blocks;
-pub(super) mod fetch_full_transaction;
 pub(super) mod fetch_taddr_transactions;
+pub(super) mod full_transactions_processor;
 pub(super) mod sync_status;
 pub(super) mod syncdata;
 pub(super) mod trial_decryptions;
